+++ packages/plugin-packages/mashroom-portal-remote-app-registry-k8s/src/js/context.ts
@@ -1,0 +1,29 @@
+// @flow
+
+import KubernetesServiceRegistry from './registry/KubernetesServiceRegistry';
+
+import {Context} from '../../type-definitions';
+
+export const _registry = new KubernetesServiceRegistry();
+let _error: string | null = null;
+let _lastScan = 0;
+
+const context: Context = {
+    get registry() {
+        return _registry;
+    },
+    get error() {
+        return _error;
+    },
+    set error(error: string | null) {
+        _error = error;
+    },
+    get lastScan() {
+        return _lastScan;
+    },
+    set lastScan(lastScan: number) {
+        _lastScan = lastScan;
+    }
+};
+
+export default context;