+++ packages/plugin-packages/mashroom-portal-remote-app-registry-k8s/src/js/webapp/mashroom-bootstrap-webapp.ts
@@ -1,0 +1,11 @@
+// @flow
+
+import webapp from './webapp';
+
+import {MashroomWebAppPluginBootstrapFunction} from "@mashroom/mashroom/type-definitions";
+
+const bootstrap: MashroomWebAppPluginBootstrapFunction = async (pluginName, pluginConfig, contextHolder) => {
+    return webapp;
+};
+
+export default bootstrap;