+++ packages/plugin-packages/mashroom-portal-remote-app-registry-k8s/src/js/webapp/webapp.ts
@@ -1,0 +1,18 @@
+
+import path from 'path';
+import express from 'express';
+import exphbs from 'express-handlebars';
+import services from './routes/services';
+
+const app = express();
+
+app.engine('handlebars', exphbs({
+}));
+app.set('view engine', 'handlebars');
+app.set('views', path.resolve(__dirname, '../../views'));
+
+app.use(express.static(path.resolve(__dirname, '../../public')));
+
+app.get('/', services);
+
+export default app;